+++ core/src/main/java/org/bookwoori/core/domain/member/controller/AuthController.java
@@ -1,16 +1,13 @@
 package org.bookwoori.core.domain.member.controller;
 
-import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.bookwoori.core.domain.member.dto.LoginResponse;
+import org.bookwoori.core.domain.member.dto.response.LoginResponse;
 import org.bookwoori.core.domain.member.service.MemberService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
